refactor(agent_backends): extract discovery submodule
Move per-backend model-list discovery + connectivity testing into
discovery.rs:

- discover_models dispatcher
- discover_openai_api_models, discover_lm_studio_models,
  discover_pi_models (pi-sdk only)
- discover_codex_models, discover_codex_native_models
- codex_models_from_debug_catalog,
  codex_native_models_from_app_server,
  models_backend_default_rank
- lm_studio_models_from_v0, filter_openai_models,
  is_openai_text_model, models_from_openai_shape
- openai_api_url (shared with the gateway translation layer that
  still lives in the facade)
- Codex CLI helpers: start_codex_native_control_session,
  codex_login_status, codex_cli_command,
  ensure_codex_native_authenticated
- test_backend_connectivity, test_codex_native_connectivity

auto_detect.rs's stale `super::{codex_cli_command,
discover_codex_models, discover_models}` import is repointed at
`super::discovery::*`. mod.rs imports the four items it actually
needs from production scope (codex_cli_command for launch_codex_login,
discover_models for refresh_agent_backend_models, openai_api_url for
the gateway's call_openai_responses helper that hasn't moved yet,
and test_backend_connectivity for the test command).

The Codex-CLI source-introspection tripwire test follows
`codex_cli_command` into discovery.rs in its current spot inside
`mod tests` by switching its `include_str!("mod.rs")` to
`include_str!("discovery.rs")` and adapting the sentinel substring
to the post-extraction layout.

Test colocation will be tightened in the facade-trim commit. Tests
covering discovery helpers stay in `mod tests` for now and reach the
pub(super) functions via `use super::discovery::{...};`.

cargo fmt --check, cargo clippy on the three CI crates with
-Dwarnings, and the agent_backends tests (111 passed) all pass. No
public-surface or wire-shape change.
